Renaming dependencies in dataflowimport files
