+++ bigtable-dataflow-parent/bigtable-dataflow-import/src/test/java/com/google/cloud/bigtable/dataflowimport/testing/BigtableTableUtils.java
@@ -15,7 +15,7 @@
  */
 package com.google.cloud.bigtable.dataflowimport.testing;
 
-import com.google.bigtable.repackaged.com.google.cloud.hbase.BigtableConfiguration;
+import com.google.bigtable.repackaged.com.google.cloud.bigtable.hbase.BigtableConfiguration;
 import com.google.cloud.bigtable.dataflow.CloudBigtableOptions;
 import com.google.common.collect.Sets;
